+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 0.2.2 (2021-06-XX)
 ------------------
 
+* Set alpha parameter as predict argument, with None as default value
 * Switch to github actions for continuous integration of the code
 * Add image explaining MAPIE internals on the README